# Remove template override