After the discussion with @eregon in #87, I think it'd be best to remove this override. I don’t think it’s worth the cost of performance and the extra complexity. Now, if you want a `<template>` tag, you need to use the `template_tag` method.

I also updated `Component#render` that was depending on this override unnecessarily. We actually don't need to respond to `render` to provide an HTML-safe render, as Rails will always call `render_in` if it's available. See #82 for more work on this where I’m trying to pass the view_context around so we can access the current request.

### Performance

Before
```
               phlex:     6919.5 i/s
      view_component:     5304.7 i/s - 1.30x  (± 0.00) slower
               cells:     3413.6 i/s - 2.03x  (± 0.00) slower
            partials:     2433.1 i/s - 2.84x  (± 0.00) slower
            dry_view:      337.7 i/s - 20.49x  (± 0.00) slower
```

After
```
               phlex:     8061.7 i/s
      view_component:     5290.0 i/s - 1.52x  (± 0.00) slower
               cells:     3442.0 i/s - 2.34x  (± 0.00) slower
            partials:     2428.9 i/s - 3.32x  (± 0.00) slower
            dry_view:      338.0 i/s - 23.85x  (± 0.00) slower
```