@jaredcwhite 2022-08-12
That seems like a reasonable compromise. The frequency with which people will be adding literal template elements into their components is likely minimal. And look at that speed gain! 😲