@joeldrapper 2022-08-12
Alright, no strong objections so far. I’m going to ship this but of course we can always re-visit if we come up with a better way forward.